Lexer's no longer cause compilation issues

+++ custom/ext/4coder_terickson_todo.cpp
@@ -14,7 +14,7 @@ struct Comment_Note
     ID_Pos_Jump_Location location;
     i64 todo_buf_title_pos;
     i64 todo_buf_text_pos;
-    
+
     Comment_Note *next;
 };
 
@@ -22,14 +22,14 @@ struct Comment_Note_File
 {
     Comment_Note *first;
     Comment_Note *last;
-    
+
     Buffer_ID buffer;
     i64 count;
     b32 finished;
     b32 updated;
-    
+
     Range_i64 out_buffer_range;
-    
+
     Arena *arena;
     Comment_Note_File *next;
 };
@@ -101,7 +101,7 @@ function void todo_handle_comment(Application_Links *app, Arena *arena, Code_Ind
         file->arena = file_arena;
         set_comment_note_file(&todo_note_table, index->buffer, file);
     }
-    
+
     String_Const_u8 TODO_STR = SCu8("@todo");
     String_Const_u8 HACK_STR = SCu8("@hack");
     String_Const_u8 str = string_substring(contents, Ii64(token));
@@ -110,33 +110,33 @@ function void todo_handle_comment(Application_Links *app, Arena *arena, Code_Ind
     {
         Comment_Note note = {};
         i64 idx = string_find_first(str, TODO_STR, StringMatch_CaseInsensitive);
-        if (idx != str.size)
+        if (idx != (i64)str.size)
         {
             note.kind = CommentNote_Todo;
             note.title = push_string_copy(&tc_global_arena, string_substring(str, Ii64(idx, idx+TODO_STR.size)));
             note.location = {index->buffer, token->pos+start_idx+idx};
             start_idx += idx+TODO_STR.size;
-            
+
             goto found;
         }
         idx = string_find_first(str, HACK_STR, StringMatch_CaseInsensitive);
-        if (idx != str.size)
+        if (idx != (i64)str.size)
         {
             note.kind = CommentNote_Hack;
             note.title = push_string_copy(&tc_global_arena, string_substring(str, Ii64(idx, idx+HACK_STR.size)));
             note.location = {index->buffer, token->pos+start_idx+idx};
             start_idx += idx+HACK_STR.size;
-            
+
             goto found;
         }
         break;
-        
-        
-        
+
+
+
         found:
-        
+
         idx = string_find_first(str, idx, ':');
-        if (idx == str.size)
+        if (idx == (i64)str.size)
         {
             push_comment_note(file, note, file->arena);
             break;
@@ -168,7 +168,7 @@ function void write_todos_to_buffer(Application_Links *app, Buffer_ID buffer, Co
         file->updated = false;
         buffer_replace_range(app, buffer, file->out_buffer_range, string_u8_empty);
         file->out_buffer_range = Ii64(idx);
-        
+
         for (Comment_Note *note = file->first;
              note != 0;
              note = note->next)
@@ -180,14 +180,14 @@ function void write_todos_to_buffer(Application_Links *app, Buffer_ID buffer, Co
                 buffer_replace_range(app, buffer, Ii64(idx), SCu8("TODO: "));
                 idx += 6;
                 break;
-                
+
                 case CommentNote_Hack:
                 buffer_replace_range(app, buffer, Ii64(idx), SCu8("HACK: "));
                 idx += 6;
                 break;
-                
+
                 default: continue;
-                
+
             }
             note->todo_buf_text_pos = idx;
             buffer_replace_range(app, buffer, Ii64(idx), note->text);
@@ -200,14 +200,14 @@ function void write_todos_to_buffer(Application_Links *app, Buffer_ID buffer, Co
         }
         file->out_buffer_range.max = idx-1;
     }
-    
+
 }
 
 global bool todo_window_open = false;
 function void todo_list_render(Application_Links *app, Comment_Note_Table *table, Rect_f32 rect)
 {
     if (!todo_window_open) return;
-    
+
     ARGB_Color background_color = fcolor_resolve(fcolor_id(defcolor_back));
     ARGB_Color border_color = fcolor_resolve(fcolor_id(defcolor_margin_active));
     ARGB_Color foreground_color = fcolor_resolve(fcolor_id(defcolor_text_default));
@@ -220,9 +220,9 @@ function void todo_list_render(Application_Links *app, Comment_Note_Table *table
         desc.parameters.pt_size = 12;
         todo_face = try_create_new_face(app, &desc);
     }
-    
+
     Scratch_Block scratch(app);
-    
+
     Buffer_ID buffer = get_buffer_by_name(app, SCu8("*TODO*"), Access_Always);
     if (!buffer_exists(app, buffer))
     {
@@ -231,13 +231,13 @@ function void todo_list_render(Application_Links *app, Comment_Note_Table *table
         buffer_set_setting(app, buffer, BufferSetting_ReadOnly, true);
     }
     write_todos_to_buffer(app, buffer, &todo_note_table);
-    
+
     draw_rectangle(app, rect, 4.f, background_color);
     draw_rectangle_outline(app, rect, 4.f, 3.f, border_color);
     rect.x0 += 3; rect.x1 -= 6;
     rect.y0 += 3; rect.y1 -= 6;
     Text_Layout_ID text_layout = text_layout_create(app, buffer, rect, {0, 0});
-    
+
     Face_ID old_face = get_face_id(app, buffer);
     buffer_set_face(app, buffer, todo_face);
     Rect_f32 old_clip = draw_set_clip(app, rect);
@@ -247,7 +247,7 @@ function void todo_list_render(Application_Links *app, Comment_Note_Table *table
     {
         Comment_Note_File *file = get_comment_note_file(&todo_note_table, buf);
         if (!file) continue;
-        
+
         for (Comment_Note *note = file->first;
              note != 0;
              note = note->next)
@@ -256,7 +256,7 @@ function void todo_list_render(Application_Links *app, Comment_Note_Table *table
             paint_text_color(app, text_layout, Ii64_size(note->todo_buf_text_pos, note->text.size), foreground_color);
         }
     }
-    
+
     draw_text_layout_default(app, text_layout);
     draw_set_clip(app, old_clip);
     buffer_set_face(app, buffer, old_face);
@@ -272,19 +272,19 @@ CUSTOM_UI_COMMAND_SIG(jump_to_note)
 CUSTOM_DOC("List all comment notes and jump to one chosen by the user.")
 {
     char *query = "Note:";
-    
+
     Scratch_Block scratch(app);
     Lister_Block lister(app, scratch);
     lister_set_query(lister, query);
     lister_set_default_handlers(lister);
-    
+
     for (Buffer_ID buffer = get_buffer_next(app, 0, Access_Always);
          buffer != 0;
          buffer = get_buffer_next(app, buffer, Access_Always))
     {
         Comment_Note_File *file = get_comment_note_file(&todo_note_table, buffer);
         if (!file) continue;
-        
+
         for (Comment_Note *note = file->first;
              note != 0;
              note = note->next)
@@ -292,7 +292,7 @@ CUSTOM_DOC("List all comment notes and jump to one chosen by the user.")
             Tiny_Jump *jump = push_array(scratch, Tiny_Jump, 1);
             jump->buffer = buffer;
             jump->pos = note->location.pos;
-            
+
             String_Const_u8 sort = {};
             switch (note->kind){
                 case CommentNote_Todo:
@@ -311,13 +311,13 @@ CUSTOM_DOC("List all comment notes and jump to one chosen by the user.")
             lister_add_item(lister, note->text, sort, jump, 0);
         }
     }
-    
+
     Lister_Result l_result = run_lister(app, lister);
     Tiny_Jump result = {};
     if (!l_result.canceled && l_result.user_data != 0){
         block_copy_struct(&result, (Tiny_Jump*)l_result.user_data);
     }
-    
+
     if (result.buffer != 0){
         View_ID view = get_this_ctx_view(app, Access_Always);
         jump_to_location(app, view, result.buffer, result.pos);
